Fix proxy not working when using a ProtoObject as superclass
ProtoObject does not define hash method, the proxy therefore was forwarding the hash message to the wrapped object.
But the wrapped object and the proxy have swapped their references.
Therefore the pointers to the objects are now pointing to the proxy.
And the hash of the object is the hash of the proxy now and vice versa.

So the proxy must understand the hash method in order to maintain the consistency of the indexed collections.